+++ client/src/components/truck/EditTruckModal.tsx
@@ -1,8 +1,9 @@
 import styled from '@emotion/styled'
 import Alert from '@mui/material/Alert';
 import Snackbar from '@mui/material/Snackbar';
-import React from 'react';
 import { useState } from 'react';
+import React from 'react';
+import Truck from "@/models/Truck";
 
 
 const MyCenteredModalContainer  = styled.div`
@@ -44,13 +45,13 @@ const Form = styled.form`
   align-items: center;
 `
 
-const EditTruckModal = ({setShowAddNewModal, setShowSuccessSnackbar}: {setShowAddNewModal: (show: boolean) => void, setShowSuccessSnackbar: (show: boolean) => void}) => {
+const EditTruckModal = ({truck, setShowEditModal, setShowSuccessSnackbar}: {truck: Truck | undefined, setShowEditModal: (show: boolean) => void, setShowSuccessSnackbar: (show: boolean) => void}) => {
 
   const [formData, setFormData] = useState({
-    registration: '',
-    makeYear: '0',
-    reservoirCapacity: 0,
-    horsepower: 0,
+    registration: truck?.registration || '',
+    makeYear: truck?.makeYear || '',
+    reservoirCapacity: truck?.reservoirCapacity || 0,
+    horsepower: truck?.horsepower || 0,
   });
   const [open, setOpen] = React.useState(false);
   const [snackbarStatus, setSnackbarStatus] = React.useState<'error' | 'success' | 'warning'>('warning');
@@ -68,14 +69,14 @@ const EditTruckModal = ({setShowAddNewModal, setShowSuccessSnackbar}: {setShowAd
       headers: {
         'Content-Type': 'application/json',
       },
-      body: JSON.stringify({ id: null, ...formData }),
+      body: JSON.stringify({ id: truck?.id, ...formData }),
     });
 
     if (response.ok) {
       setSnackbarStatus('success')
       setSnackbarText('Truck added successfully!')
       setShowSuccessSnackbar(true)
-      setShowAddNewModal(false);
+      setShowEditModal(false);
     } else {
       setSnackbarStatus('error')
       setSnackbarText('Action was not successful!')
@@ -83,7 +84,7 @@ const EditTruckModal = ({setShowAddNewModal, setShowSuccessSnackbar}: {setShowAd
     }
   };
 
-  const handleClose = (event?: React.SyntheticEvent | Event, reason?: string) => {
+  const handleClose = (_event?: React.SyntheticEvent | Event, reason?: string) => {
     if (reason === 'clickaway') {
       return;
     }
@@ -96,9 +97,13 @@ const EditTruckModal = ({setShowAddNewModal, setShowSuccessSnackbar}: {setShowAd
     <MyCenteredModalContainer>
       <MyModal>
         <div style={{height: '100%'}}>
-          <p style={{textAlign: 'center', margin: '0.75em'}}><strong>ADD NEW TRUCK</strong></p>
+          <p style={{textAlign: 'center', margin: '0.75em'}}><strong>EDIT TRUCK</strong></p>
           <Form onSubmit={handleSubmit}>
             <div>
+              <div>
+                <label>ID: </label>
+                <input type="text" value={truck?.id} readOnly />
+              </div>
               <div>
                 <label>Registration: </label>
                 <input type="text" name="registration" value={formData.registration} onChange={handleChange} />
@@ -118,7 +123,7 @@ const EditTruckModal = ({setShowAddNewModal, setShowSuccessSnackbar}: {setShowAd
             </div>
             <ButtonContainer>
               <button type="submit" style={{ width: '100px', height: '30px' }}>Save</button>
-              <button type="button" style={{ width: '100px', height: '30px' }} onClick={() => setShowAddNewModal(false)}>Close</button>
+              <button type="button" style={{ width: '100px', height: '30px' }} onClick={() => setShowEditModal(false)}>Close</button>
             </ButtonContainer>
           </Form>
         </div>
