+++ client/src/models/Truck.tsx
